# leoz-smartlane/src/main/resources/smartlane_openapi.yaml
swagger: '2.0'
info:
  title: Smartlane OpenAPI Description
  description: "The following table contains a technical specification of all API endpoints usable for Smartlane"
  version: "1.0"
host: dispatch.smartlane.io
basePath: /api
schemes:
  - https
produces:
  - application/json
consumes:
  - application/json
securityDefinitions:
  jwt:
    type: apiKey
    name: jwt
    in: header
security:
  - jwt: []
...
paths:
...
  /driver:
    get:
      summary: Drivers
      description: |
        Returns a list of all drivers with all the information available.
      operationId: getDriver
      tags:
        - Driver
      parameters:
        - $ref: "#/parameters/filterQuery"
        - in: query
          name: includeavail
          required: false
          type: boolean
          description: |
            "Returns an additional field available (which is not stored in the database) for the referring driver if the query parameter 'includeavail' is set to 'true'. Defaults to false if parameter is not supplied.

            The result for 'available' is true if these two conditions are fulfilled for a driver:

            There is at least one time frame in the driveravailability table.

            The current time is within one of the driveravailability time frames for this driver"
        - in: query
          name: includeposition
          type: boolean
          description: |
            "Returns an additional field current_position for the driver's current position if the query parameter 'includeposition' is set to 'true'. Defaults to false if parameter is not supplied.
            Both parameters can of course be combined, e.g.: /api/driver?includeavail=true&includeposition=true"
          required: false
        - in: query
          name: driverinfo
          type: boolean
          required: false
          description: |
            "Returns additional driver information (firstname, lastname, htmlcolor) if the query parameter 'driverinfo' is set to 'true'. Defaults to false if parameter is not supplied."
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Drivers"
        403:
          $ref: "#/responses/MissingAuthorization"
        422:
          $ref: "#/responses/InvalidInput"
        default:
          $ref: "#/responses/UnexpectedError"
    post:
      summary: Driver
      description: |
        Adds a new driver to the system.
      operationId: postDriver
      tags:
        - Driver
      parameters:
        - in: body
          name: driver
          required: true
          schema:
            $ref: "#/definitions/Driver"
      responses:
        201:
          description: Created
        403:
          $ref: "#/responses/MissingAuthorization"
        422:
          $ref: "#/responses/InvalidInput"
        default:
          $ref: "#/responses/UnexpectedError"
...
definitions:
...
  Drivers:
    type: object
    properties:
      num_results:
        type: integer
        format: int32
      objects:
        type: array
        items:
          $ref: '#/definitions/Driver'
      page:
        type: integer
        format: int32
      total_pages:
        type: integer
        format: int32
  Driver:
    type: object
    properties:
      id:
        type: integer
        format: int32
      firstname:
        type: string
      lastname:
        type: string
      email:
        type: string
      company_id:
        description: reference to Company
        type: integer
        format: int32
      expires:
        type: string
        format: date-time
      active:
        type: boolean
      confirmed_at:
        type: string
        format: date-time
      usertype:
        type: string
      vehicle:
        description: reference to Vehicletype
        type: string
      mobilenr:
        type: string
      htmlcolor:
        type: string
        maxLength: 6
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string

# define reusable parameters:
parameters:
  filterQuery:
    in: query
    name: q
    type: string
    description: |
      Filter query string. Can be used to filter results. [Detailed description](/page/api-handling-extras#section-filter-queries)
...
responses:
  MissingAuthorization:
    description: A failure message caused by missing authorization (403 forbidden)
    schema:
      type: string
  InvalidInput:
    description: >-
      A failure message caused by unprocessable input (e.g. no data found
      for input parameters)
    schema:
      type: string
  UnexpectedError:
    description: Unexpected error
    schema:
      $ref: '#/definitions/Error'